// client/src/interfaces/archivos.ts
export interface ReservaCSV {
    chkWeb: string; // Columna "Chk Web"
    numeroReservas: string; // Columna "N° Reservas"
    rutClienteFacturacion: string; // Columna "RUT Cliente Fact."
    nombreClienteFacturacion: string; // Columna "Nombre Cliente a Facturar"
    conductor: string; // Columna "Conductor"
    emailContactoConductor: string; // Columna "Email de Contacto Conductor"
    telefonoContactoConductor: string; // Columna "Teléfono de Contacto Conductor"
    sucursalSalida: string; // Columna "Suc. Salida"
    categoria: string; // Columna "Categoría"
    patente: string; // Columna "Patente"
    operadorEmision: string; // Columna "Operador de Emisión"
    sucursalEmision: string; // Columna "Sucursal Emisión"
    estadoReserva: string; // Columna "Estado Reserva"
    totalNeto: string; // Columna "Total Neto"
    iva: string; // Columna "IVA"
    totalFinal: string; // Columna "Total Final"
    entregarEn: string; // Columna "Entregar en ..."
    numeroVuelo: string; // Columna "N° Vuelo"
    devolverEn: string; // Columna "Devolver en ..."
}